pyproject.toml
[tool.poetry]
name = "onsagernet"
version = "0.0.1"
description = ""
authors = ["Qianxiao Li"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.poetry.dependencies]
python = "^3.11"
pdoc = "^14.7.0"
ml-dtypes = "0.4.0"
jax = "^0.4.31"
optax = "^0.2.3"
tqdm = "^4.66.5"
datasets = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.3"

[tool.poetry.group.examples.dependencies]
scikit-learn = "^1.5.2"
diffrax = "^0.6.0"
hydra-core = "^1.3.2"
jupyter = "^1.1.1"
seaborn = "^0.13.2"
hydra-joblib-launcher = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
